% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dump_dependencies_addin}
\alias{dump_dependencies_addin}
\title{Calculates the dependencies of the current file using the dependencies query
and dumps the results into the R shell}
\usage{
dump_dependencies_addin()
}
\description{
Calculates the dependencies of the current file using the dependencies query
and dumps the results into the R shell
}
