% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{dump_reconstruct_addin}
\alias{dump_reconstruct_addin}
\title{Generates a slice for the currently highlighted variable and dumps the corresponding reconstructed code into the R shell}
\usage{
dump_reconstruct_addin()
}
\description{
Generates a slice for the currently highlighted variable and dumps the corresponding reconstructed code into the R shell
}
