% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{get_reconstruction}
\alias{get_reconstruction}
\title{Generates a slice for the given filename and criterion, code fragment and criterion, or the currently highlighted variable in the active RStudio document and returns the reconstructed code fragment.}
\usage{
get_reconstruction(
  filename = NULL,
  code = NULL,
  criterion = NULL,
  print = TRUE
)
}
\arguments{
\item{filename}{The name of the file to slice. If NULL, the passed code fragment is used.}

\item{code}{The code fragment to slice, as a character. If also NULL, the currently active document is used.}

\item{criterion}{The slicing criterion to use. Needs to be non-NULL if filename or code is provided.}

\item{print}{If TRUE, the reconstructed code is printed to the console and returned invisibly. Defaults to TRUE.}
}
\value{
The reconstructed code fragment for the generated slice.
}
\description{
Generates a slice for the given filename and criterion, code fragment and criterion, or the currently highlighted variable in the active RStudio document and returns the reconstructed code fragment.
}
