% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{get_slice}
\alias{get_slice}
\title{Generates a slice for the given filename and criterion, code fragment and criterion, or the currently highlighted variable in the active RStudio document.}
\usage{
get_slice(filename = NULL, code = NULL, criterion = NULL)
}
\arguments{
\item{filename}{The name of the file to slice. If NULL, the passed code fragment is used.}

\item{code}{The code fragment to slice, as a character. If also NULL, the currently active document is used.}

\item{criterion}{The slicing criterion to use. Needs to be non-NULL if filename or code is provided.}
}
\value{
A list containing the filename, criterion, result of the slice request, a mapping from IDs to locations, and the slice locations.
}
\description{
Generates a slice for the given filename and criterion, code fragment and criterion, or the currently highlighted variable in the active RStudio document.
}
