% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{install_node_addin}
\alias{install_node_addin}
\title{Installs Node.js locally in the addin's package directory, as well as flowR's NPM package which provides a local version of the flowR shell}
\usage{
install_node_addin()
}
\description{
Installs Node.js locally in the addin's package directory, as well as flowR's NPM package which provides a local version of the flowR shell
}
