% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefs.R
\name{open_prefs_addin}
\alias{open_prefs_addin}
\title{Opens the Preferences menu, where flowR-specific settings can be changed}
\usage{
open_prefs_addin()
}
\description{
Opens the Preferences menu, where flowR-specific settings can be changed
}
