% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{show_dependencies_addin}
\alias{show_dependencies_addin}
\title{Calculates the dependencies of the current file using the dependencies query
and displays the results semi-beautifully}
\usage{
show_dependencies_addin()
}
\description{
Calculates the dependencies of the current file using the dependencies query
and displays the results semi-beautifully
}
